🎊 아두이노 리모델링 <2> 🎉

+++ display/Hope_Position.py
@@ -8,14 +8,14 @@ class position_hope:
         self.main = main
         self.ui = ui
 
-    def set_x_position(self, data: int):  # data값은 이제 계산한 값을 가지고 있어야 함.
+    def set_x_position(self, data: float):  # data값은 이제 계산한 값을 가지고 있어야 함.
         self.ui.position_x.setNum(data)
 
-    def set_y_position(self, data: int):
+    def set_y_position(self, data: float):
         self.ui.position_y.setNum(data)
 
-    def set_z_position(self, data: int):
+    def set_z_position(self, data: float):
         self.ui.position_z.setNum(data)
 
-    def set_a_position(self, data: int):
+    def set_a_position(self, data: float):
         self.ui.position_a.setNum(data)
